chore: Refactor ORD document generation
