--- __tests__/unit/version-suffix-extraction.test.js
@@ -1,397 +1,0 @@
-const { createAPIResourceTemplate } = require("../../lib/templates");
-const { DATA_PRODUCT_ANNOTATION, DATA_PRODUCT_SIMPLE_ANNOTATION, DATA_PRODUCT_TYPE, ORD_ACCESS_STRATEGY } = require("../../lib/constants");
-
-describe("Version Suffix Extraction for Data Product Services", () => {
-    const mockAppConfig = {
-        ordNamespace: "sap.test",
-        lastUpdate: "2024-01-01T00:00:00Z",
-        env: {
-            defaultVisibility: "public",
-        },
-    };
-
-    const mockPackageIds = ["sap.test:package:test:v1"];
-    const mockAccessStrategies = [{ type: ORD_ACCESS_STRATEGY.Open }];
-
-    describe("Positive Test Cases - Valid v<number> patterns", () => {
-        test("should handle .v0 suffix correctly", () => {
-            const serviceDefinition = {
-                name: "DataService.v0",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v0");
-            expect(result[0].version).toBe("0.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-
-        test("should handle .v1 suffix correctly", () => {
-            const serviceDefinition = {
-                name: "DataService.v1",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-
-        test("should handle .v2 suffix correctly", () => {
-            const serviceDefinition = {
-                name: "DataService.v2",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v2");
-            expect(result[0].version).toBe("2.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-
-        test("should handle .v10 suffix correctly", () => {
-            const serviceDefinition = {
-                name: "DataService.v10",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v10");
-            expect(result[0].version).toBe("10.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-
-        test("should handle complex service names with .v1 suffix", () => {
-            const serviceDefinition = {
-                name: "complex.DataProductService.v1",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:complex.DataProductService:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:complex.DataProductService");
-        });
-    });
-
-    describe("Negative Test Cases - Invalid patterns", () => {
-        test("should use current behavior for .v1.1 suffix (invalid pattern)", () => {
-            const serviceDefinition = {
-                name: "DataService.v1.1",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService.v1.1:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService.v1.1");
-        });
-
-        test("should use current behavior for .v1.0 suffix (invalid pattern)", () => {
-            const serviceDefinition = {
-                name: "DataService.v1.0",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService.v1.0:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService.v1.0");
-        });
-
-        test("should use current behavior for .version1 suffix (invalid pattern)", () => {
-            const serviceDefinition = {
-                name: "DataService.version1",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService.version1:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService.version1");
-        });
-
-        test("should use current behavior for .beta suffix (invalid pattern)", () => {
-            const serviceDefinition = {
-                name: "DataService.beta",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService.beta:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService.beta");
-        });
-
-        test("should use current behavior for .v suffix (invalid pattern)", () => {
-            const serviceDefinition = {
-                name: "DataService.v",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService.v:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService.v");
-        });
-    });
-
-    describe("Edge Case Tests", () => {
-        test("should use current behavior for data product service without version suffix", () => {
-            const serviceDefinition = {
-                name: "DataService",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-
-        test("should use current behavior for non-data product service with valid version suffix", () => {
-            const serviceDefinition = {
-                name: "RegularService.v2",
-                kind: 'service'
-                // No DATA_PRODUCT_ANNOTATION - this is a regular service
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:RegularService.v2:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:RegularService.v2");
-        });
-
-        test("should use current behavior for non-primary data product service", () => {
-            const serviceDefinition = {
-                name: "DataService.v2",
-                kind: 'service',
-                [DATA_PRODUCT_ANNOTATION]: "secondary", // Not primary
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService.v2:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService.v2");
-        });
-    });
-
-    describe("Data Product Specific Properties", () => {
-        test("should maintain data product specific properties with version extraction", () => {
-            const serviceDefinition = {
-                name: "DataService.v2",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].direction).toBe("outbound");
-            expect(result[0].apiProtocol).toBe("sap.dp:data-subscription-api:v1");
-            expect(result[0].entryPoints).toEqual([]);
-            expect(result[0].resourceDefinitions).toHaveLength(1);
-            expect(result[0].resourceDefinitions[0].type).toBe("sap-csn-interop-effective-v1");
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v2");
-            expect(result[0].version).toBe("2.0.0");
-        });
-    });
-
-    describe("Tests with @data.product annotation", () => {
-        test("should handle .v1 suffix correctly with @data.product annotation", () => {
-            const serviceDefinition = {
-                name: "DataService.v1",
-                [DATA_PRODUCT_SIMPLE_ANNOTATION]: true,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-            expect(result[0].direction).toBe("outbound");
-            expect(result[0].apiProtocol).toBe("sap.dp:data-subscription-api:v1");
-        });
-
-        test("should handle .v2 suffix correctly with @data.product annotation", () => {
-            const serviceDefinition = {
-                name: "DataService.v2",
-                [DATA_PRODUCT_SIMPLE_ANNOTATION]: "yes",
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v2");
-            expect(result[0].version).toBe("2.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-
-        test("should use current behavior when @data.product is false", () => {
-            const serviceDefinition = {
-                name: "DataService.v2",
-                kind: 'service',
-                [DATA_PRODUCT_SIMPLE_ANNOTATION]: false,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService.v2:v1");
-            expect(result[0].version).toBe("1.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService.v2");
-        });
-
-        test("should prioritize @DataIntegration.dataProduct.type over @data.product for version extraction", () => {
-            const serviceDefinition = {
-                name: "DataService.v3",
-                [DATA_PRODUCT_ANNOTATION]: DATA_PRODUCT_TYPE.primary,
-                [DATA_PRODUCT_SIMPLE_ANNOTATION]: false,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v3");
-            expect(result[0].version).toBe("3.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-
-        test("should apply @data.product when @DataIntegration.dataProduct.type is not primary", () => {
-            const serviceDefinition = {
-                name: "DataService.v2",
-                [DATA_PRODUCT_ANNOTATION]: "secondary",
-                [DATA_PRODUCT_SIMPLE_ANNOTATION]: true,
-            };
-
-            const result = createAPIResourceTemplate(
-                serviceDefinition,
-                mockAppConfig,
-                mockPackageIds,
-                mockAccessStrategies,
-            );
-
-            expect(result).toHaveLength(1);
-            expect(result[0].ordId).toBe("sap.test:apiResource:DataService:v2");
-            expect(result[0].version).toBe("2.0.0");
-            expect(result[0].partOfGroups[0]).toBe("sap.cds:service:sap.test:DataService");
-        });
-    });
-});
