+++ lib/common/utils.js
@@ -1,14 +1,19 @@
 const {
+    ENTITY_RELATIONSHIP_ANNOTATION,
     DATA_PRODUCT_ANNOTATION,
     DATA_PRODUCT_TYPE,
     DATA_PRODUCT_SIMPLE_ANNOTATION,
     ORD_EXTENSIONS_PREFIX,
     RESOURCE_VISIBILITY,
     ALLOWED_VISIBILITY,
-    SUPPORTED_IMPLEMENTATIONSTANDARD_VERSIONS, EXTERNAL_SERVICE_ANNOTATION, EXTERNAL_DP_ORD_ID_ANNOTATION,
+    SUPPORTED_IMPLEMENTATIONSTANDARD_VERSIONS,
+    EXTERNAL_SERVICE_ANNOTATION,
+    EXTERNAL_DP_ORD_ID_ANNOTATION,
+    CDS_ELEMENT_KIND,
 } = require("../constants");
 const _ = require("lodash");
 const Logger = require("../logger");
+const cds = require("@sap/cds");
 
 function getRFC3339Date() {
     const now = new Date();
@@ -94,13 +99,51 @@ function isExternalDataProduct(definition) {
     );
 }
 
+function isExposedEntityType(definition) {
+    return (
+        !definition.name.includes(".texts") &&
+        definition[ENTITY_RELATIONSHIP_ANNOTATION] &&
+        definition.kind === CDS_ELEMENT_KIND.entity &&
+        definition["_service"]?.["@protocol"] !== "none" &&
+        !isBlockedServiceName(definition["_service"]?.name)
+    );
+}
+
+function isBlockedServiceName(name) {
+    return ["cds.xt.MTXServices", "MtxOrdProviderService", "OpenResourceDiscoveryService"] //
+        .some((blocked) => name?.includes(blocked));
+}
+
+function isValidService(definition) {
+    const isExternalService =
+        Object.keys(cds).includes("requires") && Object.keys(cds.requires).includes(definition.name);
+
+    return (
+        !isExternalService &&
+        !definition["@cds.external"] &&
+        definition["@protocol"] !== "none" &&
+        !isBlockedServiceName(definition.name) &&
+        definition.kind === CDS_ELEMENT_KIND.service
+    );
+}
+
+function isEventsOnlyService(definition) {
+    return (
+        Object.keys(definition.events || {}).length > 0 &&
+        ["actions", "entities", "functions"].every((key) => Object.keys(definition[key] || {}).length === 0)
+    );
+}
 
 module.exports = {
+    isValidService,
     getRFC3339Date,
     readORDExtensions,
     resolveVisibility,
     flattenEntityGraph,
     resolveServiceName,
+    isExposedEntityType,
+    isEventsOnlyService,
+    isBlockedServiceName,
     isExternalDataProduct,
     isPrimaryDataProductService,
 };
