+++ lib/configuration.js
@@ -1,14 +1,12 @@
 const path = require("path");
 const cds = require("@sap/cds");
-const _ = require("lodash");
 const { readFileSync } = require("fs");
 
 const Logger = require("./logger");
 const defaults = require("./defaults");
+const { getRFC3339Date } = require("./common/utils");
 const { createAuthConfig } = require("./auth/authentication");
-const { createEntityTypeMappingsItemTemplate } = require("./templates");
-const { getRFC3339Date, isExternalDataProduct } = require("./common/utils");
-const { CONTENT_MERGE_KEY, CDS_ELEMENT_KIND, EXTERNAL_DP_ORD_ID_ANNOTATION } = require("./constants");
+const { CDS_ELEMENT_KIND, ORD_EXTENSIONS_PREFIX } = require("./constants");
 
 module.exports = class Configuration {
     constructor(csn) {
@@ -31,33 +29,27 @@ module.exports = class Configuration {
         this._packageName = packageName;
         this._ordNamespace = ordNamespace;
         this._lastUpdate = getRFC3339Date();
+        this._csn = this._propagateORDVisibility(csn);
         this._internalNamespace = env?.internalNamespace;
-        this._serviceNames = this._resolveServiceNames(csn);
         this._existingProductORDId = env?.existingProductORDId;
-        this._apiResourceNames = this._resolveApiResourceNames(csn);
-        this._entityTypeTargets = this._resolveEntityTypeTargets(csn);
-        this._eventServiceNames = this._resolveEventServiceNames(csn);
         this._appName = packageName.replace(/^@/, "").replace(/[@/]/g, "-");
-        this._externalServiceNames = this._resolveExternalServiceNames(csn);
         this._policyLevels = env?.policyLevels || (env?.policyLevel && [env?.policyLevel]) || defaults.policyLevels;
-    }
-
-    _resolveExternalServiceNames(csn) {
-        return Object.keys(csn.definitions)
-            .filter((name) => isExternalDataProduct(csn.definitions[name]))
-            .filter((name) => "apiResource" === csn.definitions[name][EXTERNAL_DP_ORD_ID_ANNOTATION].split(":")[1]);
-    }
-
-    get authConfig() {
-        return this._authConfig;
     }
 
     get env() {
         return this._env;
     }
 
+    get csn() {
+        return this._csn;
+    }
+
     get appName() {
         return this._appName;
+    }
+
+    get authConfig() {
+        return this._authConfig;
     }
 
     get lastUpdate() {
@@ -68,36 +60,16 @@ module.exports = class Configuration {
         return this._packageName;
     }
 
-    get serviceNames() {
-        return this._serviceNames;
-    }
-
     get ordNamespace() {
         return this._ordNamespace;
-    }
-
-    get internalNamespace() {
-        return this._internalNamespace;
     }
 
     get policyLevels() {
         return this._policyLevels;
     }
 
-    get apiResourceNames() {
-        return this._apiResourceNames;
-    }
-
-    get externalServiceNames() {
-        return this._externalServiceNames;
-    }
-
-    get entityTypeTargets() {
-        return this._entityTypeTargets;
-    }
-
-    get eventServiceNames() {
-        return this._eventServiceNames;
+    get internalNamespace() {
+        return this._internalNamespace;
     }
 
     get hasSAPPolicyLevel() {
@@ -106,11 +78,6 @@ module.exports = class Configuration {
 
     get existingProductORDId() {
         return this._existingProductORDId;
-    }
-
-    _resolveServiceNames(csn) {
-        return Object.keys(csn.definitions) //
-            .filter((name) => this._isValidService(name, csn.definitions[name]));
     }
 
     _loadNameFromPackageJson() {
@@ -123,60 +90,22 @@ module.exports = class Configuration {
         return JSON.parse(readFileSync(packageJsonPath, "utf-8")).name;
     }
 
-    _isBlockedServiceName(name) {
-        return ["cds.xt.MTXServices", "MtxOrdProviderService", "OpenResourceDiscoveryService"] //
-            .some((blocked) => name.includes(blocked));
-    }
+    _propagateORDVisibility(csn) {
+        for (const [name, def] of Object.entries(csn.definitions)) {
+            if (def.kind === CDS_ELEMENT_KIND.service && def[ORD_EXTENSIONS_PREFIX + "visibility"]) {
+                const serviceName = name;
+                const serviceVisibility = def[ORD_EXTENSIONS_PREFIX + "visibility"];
+                for (const serviceChildrenDef of csn.definitions) {
+                    if (
+                        serviceChildrenDef.name.startsWith(serviceName + ".") &&
+                        !serviceChildrenDef[ORD_EXTENSIONS_PREFIX + "visibility"]
+                    ) {
+                        serviceChildrenDef[ORD_EXTENSIONS_PREFIX + "visibility"] = serviceVisibility;
+                    }
+                }
+            }
+        }
 
-    _resolveApiResourceNames(csn) {
-        return Object.keys(csn.definitions)
-            .filter((name) => this._isValidService(name, csn.definitions[name]))
-            .filter((name) => !this._serviceOnlyContainsEvents(csn.definitions[name]));
-    }
-
-    _resolveEntityTypeTargets(csn) {
-        return _.uniqBy(
-            Object.keys(csn.definitions)
-                .filter((name) => !name.includes(".texts"))
-                .filter((name) => !this._isBlockedServiceName(name))
-                .filter((name) => csn.definitions[name].kind === CDS_ELEMENT_KIND.entity)
-                .filter((name) => csn.definitions[name]["_service"]?.["@protocol"] !== "none")
-                .flatMap((name) => createEntityTypeMappingsItemTemplate(csn.definitions[name]) || []),
-            CONTENT_MERGE_KEY,
-        );
-    }
-
-    _resolveEventServiceNames(csn) {
-        const serviceNames = this._resolveServiceNames(csn);
-
-        return [
-            ...new Set(
-                Object.keys(csn.definitions)
-                    .filter((name) => !this._isBlockedServiceName(name))
-                    .filter((name) => csn.definitions[name].kind === CDS_ELEMENT_KIND.event)
-                    .filter((name) => csn.definitions[name]["_service"]?.["@protocol"] !== "none")
-                    .filter((name) => serviceNames.some((serviceName) => name.startsWith(`${serviceName}.`)))
-                    .map((name) => serviceNames.find((serviceName) => name.startsWith(`${serviceName}.`))),
-            ),
-        ];
-    }
-
-    _isValidService(key, definition) {
-        const isExternalService = Object.keys(cds).includes("requires") && Object.keys(cds.requires).includes(key);
-
-        return (
-            definition.kind === CDS_ELEMENT_KIND.service &&
-            !definition["@cds.external"] &&
-            definition["@protocol"] !== "none" &&
-            !isExternalService &&
-            !this._isBlockedServiceName(key)
-        );
-    }
-
-    _serviceOnlyContainsEvents(definition) {
-        return (
-            Object.keys(definition.events || {}).length > 0 &&
-            ["actions", "entities", "functions"].every((key) => Object.keys(definition[key] || {}).length === 0)
-        );
+        return csn;
     }
 };
