+++ lib/ord.js
@@ -3,40 +3,24 @@ const cds = require("@sap/cds");
 
 const Logger = require("./logger");
 const defaults = require("./defaults");
-const { resolveVisibility } = require("./common/utils");
 const Configuration = require("./configuration");
-const { RESOURCE_VISIBILITY } = require("./constants");
 const { createGroupsTemplateForService } = require("./templates/group");
 const { createEntityTypeTemplate } = require("./templates/entity-type");
-const { createIntegrationDependency } = require("./templates/integration-dependency");
-const { getAccessStrategiesFromAuthConfig } = require("./access-strategies");
+const { createAPIResourceTemplate } = require("./templates/api-resource");
 const { createEventResourceTemplate } = require("./templates/event-resource");
-const { createAPIResourceTemplate, _propagateORDVisibility } = require("./templates");
+const { createIntegrationDependency } = require("./templates/integration-dependency");
+const { isValidService, resolveVisibility, isEventsOnlyService, isExposedEntityType } = require("./common/utils");
+const { CONTENT_MERGE_KEY, RESOURCE_VISIBILITY, CDS_ELEMENT_KIND } = require("./constants");
 const {
     getCustomORDContent,
     compareAndHandleCustomORDContentWithExistingContent,
 } = require("./extend-ord-with-custom");
 
-const _getGroups = (csn, appConfig) => {
-    return appConfig.serviceNames
-        .map((name) => csn.definitions[name])
-        .filter((srvDefinition) => resolveVisibility(appConfig, srvDefinition) !== RESOURCE_VISIBILITY.private)
-        .flatMap((srvDefinition) => createGroupsTemplateForService(srvDefinition, appConfig));
-};
-
-const _getAPIResources = (csn, appConfig, packageIds) => {
-    const accessStrategies = getAccessStrategiesFromAuthConfig(appConfig.authConfig.accessStrategies || []);
-
-    return appConfig.apiResourceNames
-        .map((name) => csn.definitions[name])
-        .flatMap((srvDefinition) => createAPIResourceTemplate(srvDefinition, appConfig, packageIds, accessStrategies));
-};
-
-const _getEventResources = (csn, appConfig) => {
-    return appConfig.eventServiceNames
-        .map((name) => csn.definitions[name])
-        .map((srvDefinition) => createEventResourceTemplate(srvDefinition, appConfig))
-        .filter((resource) => resource.visibility !== RESOURCE_VISIBILITY.private);
+const _getGroups = (appConfig) => {
+    return Object.values(appConfig.csn.definitions)
+        .filter((definition) => isValidService(definition))
+        .filter((service) => resolveVisibility(appConfig, service) !== RESOURCE_VISIBILITY.private)
+        .flatMap((service) => createGroupsTemplateForService(service, appConfig));
 };
 
 const _getProducts = (appConfig) => {
@@ -59,42 +43,37 @@ const _getProducts = (appConfig) => {
 const _getEntityTypes = (appConfig) => {
     return appConfig.hasSAPPolicyLevel
         ? [] // If SAP policy level is present, don't create entity type, they must be in the central repository
-        : appConfig.entityTypeTargets
-              .filter((entity) => !entity.isODMMapping)
-              .map((entity) => createEntityTypeTemplate(appConfig, entity))
-              .filter((entity) => entity.visibility !== RESOURCE_VISIBILITY.private);
+        : _.uniqBy(
+              Object.values(appConfig.csn.definitions)
+                  .filter((definition) => isExposedEntityType(definition))
+                  .map((entity) => createEntityTypeTemplate(appConfig, entity))
+                  .filter((entity) => entity.visibility !== RESOURCE_VISIBILITY.private),
+              CONTENT_MERGE_KEY,
+          );
 };
 
-const _getIntegrationDependencies = (csn, appConfig) => {
-    return !appConfig.externalServiceNames?.length ? [] : [createIntegrationDependency(csn, appConfig)];
+const _getAPIResources = (appConfig) => {
+    return Object.values(appConfig.csn.definitions)
+        .filter((definition) => isValidService(definition) && !isEventsOnlyService(definition))
+        .flatMap((service) => createAPIResourceTemplate(service, appConfig));
 };
 
-const _createDefaultORDDocument = (linkedCsn, appConfig) => {
-    const products = _getProducts(appConfig);
-    const packages = defaults.packages(appConfig, products);
-    const packageIds = packages?.map((pkg) => pkg.ordId) || [];
-    const entityTypes = _getEntityTypes(appConfig);
-    const eventResources = _getEventResources(linkedCsn, appConfig);
-    const apiResources = _getAPIResources(linkedCsn, appConfig, packageIds);
-    const integrationDependencies = _getIntegrationDependencies(linkedCsn, appConfig);
+const _getEventResources = (appConfig) => {
+    const services = _.uniq(
+        Object.values(appConfig.csn.definitions)
+            .filter((definition) => definition.kind === CDS_ELEMENT_KIND.event)
+            .filter((definition) => isValidService(definition["_service"]))
+            .map((definition) => definition["_service"]),
+    );
 
-    return {
-        // Unconditionally added top-level properties
-        $schema: "https://open-resource-discovery.github.io/specification/spec-v1/interfaces/Document.schema.json",
-        policyLevels: appConfig.policyLevels,
-        packages: packages,
-        description: appConfig.env?.description || defaults.description,
-        openResourceDiscovery: appConfig.env?.openResourceDiscovery || defaults.openResourceDiscovery,
+    return services
+        .map((service) => createEventResourceTemplate(service, appConfig))
+        .filter((resource) => resource.visibility !== RESOURCE_VISIBILITY.private);
+};
 
-        // Conditionally added top-level properties
-        ...(!entityTypes.length ? {} : { entityTypes: entityTypes }),
-        ...(!apiResources.length ? {} : { apiResources: apiResources }),
-        ...(!eventResources.length ? {} : { eventResources: eventResources }),
-        ...(appConfig.existingProductORDId ? {} : { products: [products[0]] }),
-        ...(!appConfig.serviceNames.length ? {} : { groups: _getGroups(linkedCsn, appConfig) }),
-        ...(!integrationDependencies.length ? {} : { integrationDependencies: integrationDependencies }),
-        ...(!appConfig.env?.consumptionBundles?.length ? {} : { consumptionBundles: appConfig.env.consumptionBundles }),
-    };
+const _getIntegrationDependencies = (appConfig) => {
+    return [createIntegrationDependency(appConfig)] //
+        .filter((dependency) => dependency.aspects?.length > 0);
 };
 
 const _filterUnusedPackages = (ordDocument) => {
@@ -116,14 +95,33 @@ const _filterUnusedPackages = (ordDocument) => {
 };
 
 module.exports = (csn, extensions = []) => {
-    const linkedCsn = _propagateORDVisibility(cds.linked(csn));
-    const appConfig = new Configuration(linkedCsn);
+    const appConfig = new Configuration(cds.linked(csn));
+    const groups = _getGroups(appConfig);
+    const products = _getProducts(appConfig);
+    const entityTypes = _getEntityTypes(appConfig);
+    const apiResources = _getAPIResources(appConfig);
+    const eventResources = _getEventResources(appConfig);
+    const consumptionBundles = appConfig.env?.consumptionBundles ?? [];
+    const integrationDependencies = _getIntegrationDependencies(appConfig);
     const ordDocument = [...(extensions || []), getCustomORDContent(appConfig)]
         .filter((extension) => !!extension)
-        .reduce(
-            (document, extension) => compareAndHandleCustomORDContentWithExistingContent(document, extension),
-            _createDefaultORDDocument(linkedCsn, appConfig),
-        );
+        .reduce((document, extension) => compareAndHandleCustomORDContentWithExistingContent(document, extension), {
+            // Unconditionally added top-level properties
+            $schema: "https://open-resource-discovery.github.io/specification/spec-v1/interfaces/Document.schema.json",
+            policyLevels: appConfig.policyLevels,
+            packages: defaults.packages(appConfig, products),
+            description: appConfig.env?.description || defaults.description,
+            openResourceDiscovery: appConfig.env?.openResourceDiscovery || defaults.openResourceDiscovery,
+
+            // Conditionally added top-level properties
+            ...(!groups.length ? {} : { groups: groups }),
+            ...(!entityTypes.length ? {} : { entityTypes: entityTypes }),
+            ...(!apiResources.length ? {} : { apiResources: apiResources }),
+            ...(!eventResources.length ? {} : { eventResources: eventResources }),
+            ...(appConfig.existingProductORDId ? {} : { products: [products[0]] }),
+            ...(!consumptionBundles.length ? {} : { consumptionBundles: consumptionBundles }),
+            ...(!integrationDependencies.length ? {} : { integrationDependencies: integrationDependencies }),
+        });
 
     return Object.assign(ordDocument, { packages: _filterUnusedPackages(ordDocument) });
 };
