+++ lib/templates/api-resource.js
@@ -1,58 +1,45 @@
-const _ = require("lodash");
-const defaults = require("./defaults");
-
-const Logger = require("./logger");
-const { ensureAccessStrategies } = require("./access-strategies");
-const { resolveApiResourceProtocol } = require("./protocol-resolver");
-const { readORDExtensions, isPrimaryDataProductService } = require("./common/utils");
-const { resolveVisibility, resolveServiceName, flattenEntityGraph } = require("./common/utils");
+const Logger = require("../logger");
+const defaults = require("../defaults");
+const { resolveApiResourceProtocol } = require("../protocol-resolver");
+const { readORDExtensions, isPrimaryDataProductService } = require("../common/utils");
+const { resolveVisibility, resolveServiceName, flattenEntityGraph } = require("../common/utils");
+const { ensureAccessStrategies, getAccessStrategiesFromAuthConfig } = require("../access-strategies");
 const {
     DESCRIPTION_PREFIX,
-    ENTITY_RELATIONSHIP_ANNOTATION,
-    ORD_EXTENSIONS_PREFIX,
-    ORD_ODM_ENTITY_NAME_ANNOTATION,
     ORD_RESOURCE_TYPE,
     RESOURCE_VISIBILITY,
     SHORT_DESCRIPTION_PREFIX,
-    CONTENT_MERGE_KEY,
-    CDS_ELEMENT_KIND,
     ORD_API_PROTOCOL,
     MCP_RESOURCE_DEFINITION_TYPE,
-} = require("./constants");
-
-/**
- * This is a template function to create item of entityTypeMappings array.
- *
- * @param {string} entity The entity definition.
- * @returns {Object} An entry of the entityTypeMappings array.
- */
-const createEntityTypeMappingsItemTemplate = (entity) => {
-    const results = [];
-    if (entity[ORD_ODM_ENTITY_NAME_ANNOTATION]) {
-        results.push({
-            ordId: `sap.odm:entityType:${entity[ORD_ODM_ENTITY_NAME_ANNOTATION]}:v1`,
-            entityName: entity[ORD_ODM_ENTITY_NAME_ANNOTATION],
-            isODMMapping: true,
-            ...entity,
-        });
-    }
-    if (entity[ENTITY_RELATIONSHIP_ANNOTATION]) {
-        const ordIdParts = entity[ENTITY_RELATIONSHIP_ANNOTATION].split(":");
-        const namespace = ordIdParts[0];
-        const entityName = ordIdParts[1];
-        const version = entity[`${ORD_EXTENSIONS_PREFIX}version`] || ordIdParts[2]?.substring(1) || "1";
-        results.push({
-            ordId: `${namespace}:entityType:${entityName}:v${version.split(".")[0]}`,
-            entityName,
-            ...entity,
-        });
-    }
-    if (results.length === 0) return;
-    return results;
-};
+    ORD_EXTENSIONS_PREFIX,
+    ORD_ODM_ENTITY_NAME_ANNOTATION,
+    ENTITY_RELATIONSHIP_ANNOTATION,
+} = require("../constants");
 
 function _getGroupID(appConfig, srvDefinition) {
     return `${defaults.groupTypeId}:${appConfig.ordNamespace}:${resolveServiceName(appConfig, srvDefinition)}`;
+}
+
+function _getExposedEntityTypes(service) {
+    const ordIds = new Set(
+        Object.values(service.entities ?? {})
+            .flatMap((entity) => flattenEntityGraph(entity))
+            .flatMap((entity) => {
+                const [namespace, name, version = "v1"] = entity[ENTITY_RELATIONSHIP_ANNOTATION]?.split(":") || [];
+                const major = (entity[`${ORD_EXTENSIONS_PREFIX}version`] ?? version.substring(1)).split(".")[0];
+
+                return [
+                    ...(!entity[ORD_ODM_ENTITY_NAME_ANNOTATION]
+                        ? []
+                        : [`sap.odm:entityType:${entity[ORD_ODM_ENTITY_NAME_ANNOTATION]}:v1`]),
+                    ...(!entity[ENTITY_RELATIONSHIP_ANNOTATION]
+                        ? []
+                        : [entity["@ORD.Extensions.ordId"] ?? `${namespace}:entityType:${name}:v${major}`]),
+                ];
+            }),
+    );
+
+    return [...ordIds].map((ordId) => ({ ordId: ordId }));
 }
 
 /**
@@ -91,16 +78,16 @@ function _getResourceDefinition(resourceType, mediaType, ordId, serviceName, fil
  * Ensures proper visibility compliance by checking associated EntityTypes.
  * @param {object} srvDefinition The definition of the service
  * @param {object} appConfig - The application configuration.
- * @param {Array} packageIds - The available package identifiers.
- * @param {Array<{type: string}>} accessStrategies The array of accessStrategies objects
  * @returns {Array} An array of objects for the API Resources.
  */
-const createAPIResourceTemplate = (srvDefinition, appConfig, packageIds, accessStrategies) => {
+const createAPIResourceTemplate = (srvDefinition, appConfig) => {
     const ordExtensions = readORDExtensions(srvDefinition);
     const visibility = resolveVisibility(appConfig, srvDefinition);
+    const packageIds = defaults.packages(appConfig)?.map((pkg) => pkg.ordId) || [];
     const packageId = _getPackageID(appConfig.ordNamespace, packageIds, ORD_RESOURCE_TYPE.api, visibility);
     const protocolResults = resolveApiResourceProtocol(srvDefinition);
     const apiResources = [];
+    const accessStrategies = getAccessStrategiesFromAuthConfig(appConfig.authConfig.accessStrategies || []);
 
     // If no protocols were generated, skip this service
     if (protocolResults.length === 0) {
@@ -228,7 +215,7 @@ const createAPIResourceTemplate = (srvDefinition, appConfig, packageIds, accessS
             extensible: {
                 supported: "no",
             },
-            ...(exposedEntityTypes ? { exposedEntityTypes } : []),
+            ...(exposedEntityTypes.length ? { exposedEntityTypes } : []),
             ...ordExtensions,
             ...protocolExtensions,
         };
@@ -249,25 +236,6 @@ const createAPIResourceTemplate = (srvDefinition, appConfig, packageIds, accessS
 
     return apiResources;
 };
-
-function _getExposedEntityTypes(definitionObj) {
-    if (!definitionObj.entities) {
-        return;
-    }
-    const entities = Object.values(definitionObj.entities).flatMap((entity) => {
-        const entityData = flattenEntityGraph(entity).flatMap(createEntityTypeMappingsItemTemplate).filter(Boolean);
-        return _.uniqBy(entityData, CONTENT_MERGE_KEY);
-    });
-    const exposedEntityTypes = _.uniqBy(entities, CONTENT_MERGE_KEY)
-        .filter((entity) => !!entity)
-        .map((entity) => ({
-            ordId: entity.isODMMapping ? entity.ordId : entity[`${ORD_EXTENSIONS_PREFIX}ordId`] || entity.ordId,
-        }));
-
-    if (exposedEntityTypes.length > 0) {
-        return exposedEntityTypes;
-    }
-}
 
 /**
  * Extracts version suffix from service name for data product services.
@@ -312,29 +280,8 @@ function _getPackageID(namespace, packageIds, resourceType, visibility = RESOURC
     return packageIds.find((id) => id.includes(`-${resourceType}-`)) || packageIds.find((id) => id.includes(namespace));
 }
 
-function _propagateORDVisibility(model) {
-    for (const [name, def] of Object.entries(model.definitions)) {
-        if (def.kind === CDS_ELEMENT_KIND.service && def[ORD_EXTENSIONS_PREFIX + "visibility"]) {
-            const serviceName = name;
-            const serviceVisibility = def[ORD_EXTENSIONS_PREFIX + "visibility"];
-            for (const serviceChildrenDef of model.definitions) {
-                if (
-                    serviceChildrenDef.name.startsWith(serviceName + ".") &&
-                    !serviceChildrenDef[ORD_EXTENSIONS_PREFIX + "visibility"]
-                ) {
-                    serviceChildrenDef[ORD_EXTENSIONS_PREFIX + "visibility"] = serviceVisibility;
-                }
-            }
-        }
-    }
-
-    return model;
-}
-
 module.exports = {
-    createEntityTypeMappingsItemTemplate,
     createAPIResourceTemplate,
     _getPackageID,
     _getExposedEntityTypes,
-    _propagateORDVisibility,
 };
