+++ lib/templates/integration-dependency.js
@@ -1,5 +1,5 @@
 const defaults = require("../defaults");
-const { readORDExtensions } = require("../common/utils");
+const { readORDExtensions, isExternalDataProduct } = require("../common/utils");
 const { RESOURCE_VISIBILITY, EXTERNAL_DP_ORD_ID_ANNOTATION } = require("../constants");
 
 const RESOLVERS = Object.freeze({
@@ -12,9 +12,10 @@ const RESOLVERS = Object.freeze({
     version: (appConfig) => {
         return appConfig.env?.integrationDependency?.version || "1.0.0";
     },
-    aspects: (appConfig, csn) => {
-        return appConfig.externalServiceNames
-            .map((name) => csn.definitions[name])
+    aspects: (appConfig) => {
+        return Object.values(appConfig.csn.definitions)
+            .filter((definition) => isExternalDataProduct(definition))
+            .filter((service) => "apiResource" === service[EXTERNAL_DP_ORD_ID_ANNOTATION].split(":")[1])
             .map((service) => {
                 const version = service[EXTERNAL_DP_ORD_ID_ANNOTATION].split(":")[3] ?? "v1";
 
@@ -53,11 +54,10 @@ const RESOLVERS = Object.freeze({
 
 /**
  * Creates a single IntegrationDependency with one aspect per external service.
- * @param {Object} csn - The CSN definitions object
  * @param {Object} appConfig - The application configuration
  * @returns {Object} IntegrationDependency object
  */
-function createIntegrationDependency(csn, appConfig) {
+function createIntegrationDependency(appConfig) {
     return {
         mandatory: false,
         releaseStatus: "active",
@@ -65,7 +65,7 @@ function createIntegrationDependency(csn, appConfig) {
 
         ordId: RESOLVERS.ordId(appConfig),
         version: RESOLVERS.version(appConfig),
-        aspects: RESOLVERS.aspects(appConfig, csn),
+        aspects: RESOLVERS.aspects(appConfig),
         visibility: RESOLVERS.visibility(appConfig),
         partOfPackage: RESOLVERS.partOfPackage(appConfig),
